Add a common handler that uses mappings to handle commands.
By having the user and group classes extend this we move the handling
into a common location that simplifies the calls, but keep the actual
mappings part of the relevant classes, where they shuold be.

Also move the validate methods into the group and user classes, which
will need implementing at some point.